pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [ "setuptools", "wheel" ]

[project]
name = "rhgardenservices"
version = "4.0.0"
description = "Corporate website for R H Garden Services"
authors = [
  { name = "Ben Gosney", email = "bengosney@googlemail.com" },
]
# [[[cog
# import cog
# with open(".python-version") as f:
#   version = ".".join(f.read().strip().split('.')[0:2])
# cog.outl(f"requires-python = \">={version}\"")
# ]]]
requires-python = ">=3.14"
# [[[end]]] (sum: Vc9kRGWPzK) (sum: Vc9kRGWPzK)

classifiers = [
  # [[[cog
  # cog.outl(f"\"Programming Language :: Python :: {version.split('.')[0]} :: Only\",")
  # cog.outl(f"\"Programming Language :: Python :: {version}\",")
  # ]]]
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.14",
  # [[[end]]] (sum: 7gn4d5+oKr) (sum: 7gn4d5+oKr)
]
dependencies = [
  "boto3",
  "dj-database-url",
  "django",
  "django-csp",
  "django-health-check",
  "django-storages",
  "gunicorn",
  "honeybadger",
  "opencv-python-headless",
  "psycopg[binary,pool]",
  "redis",
  "rustface",
  "wagtail",
  "wheel",
  "whitenoise",
  "whitenoise[brotli]",
]

optional-dependencies.dev = [
  "cogapp",
  "django-browser-reload",
  "django-debug-toolbar",
  "django-debugtools",
  "django-stubs",
  "icecream",
  "model-bakery",
  "mypy",
  "packaging",
  "pre-commit",
  "pytest",
  "pytest-clarity",
  "pytest-cov",
  "pytest-django",
  "pytest-picked",
  "pytest-watch",
  "ruff",
  "wagtail-placeholder-images",
  "wrapt",
]

[tool.ruff]
# [[[cog
# cog.outl(f"target-version = \"py{version.replace('.','')}\"")
# ]]]
target-version = "py314"
# [[[end]]] (sum: 0xk3kDfGuA) (sum: 0xk3kDfGuA)
line-length = 120
exclude = [ "migrations" ]

format.indent-style = "space"
format.quote-style = "double"
format.docstring-code-format = true

lint.select = [ "C90", "DJ", "E", "ERA", "F", "FURB", "I", "N", "PERF", "PL", "RUF", "TRY", "UP" ]
lint.ignore = [ "RUF012" ]

lint.isort.section-order = [
  "future",
  "standard-library",
  "django",
  "wagtail",
  "third-party",
  "first-party",
  "local-folder",
]

lint.isort.sections.django = [ "django" ]
lint.isort.sections.wagtail = [ "wagtail" ]

[tool.pyproject-fmt]
# [[[cog
# cog.outl(f"max_supported_python = \"{version}\"")
# ]]]
max_supported_python = "3.14"
# [[[end]]] (sum: SnZxKjKo5y) (sum: SnZxKjKo5y)
